fix(template-shifts): resolve client/server boundary violation
- Extract types (SplitDay, TemplateCastShift, TemplateShiftData) and
  generateDaysForSplit() to lib/template-shift-utils.ts (no 'use server')
- lib/actions/template-shifts.ts now exports only getTemplateShiftData()
- TemplateShiftEditor: import from lib/template-shift-utils (was server action)
- export-template-shifts/route.ts: import type from lib/template-shift-utils
- Fix implicit any params in route.ts forEach callback

Before: Build error - 'generateDaysForSplit' not found in server module
After: TS errors 0
